embed

+++ x/cronos/events/bindings/src/LLama.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.4;
 
 interface ILLamaModule {
-    function inference(string calldata prompt, uint256 seed, uint256 steps) external payable returns (string memory result);
+    function inference(string calldata prompt, int64 seed, int32 steps) external payable returns (string memory result);
 }
